requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
pydantic==2.10.3
pydantic-settings==2.7.0
google-generativeai==0.8.3
llama-index-core==0.12.8
llama-index-readers-file==0.4.2
llama-index-vector-stores-faiss==0.3.0
llama-index-embeddings-huggingface==0.5.0
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
transformers==4.41.2
tokenizers==0.19.1
huggingface-hub==0.27.1
numpy==1.26.4
faster-whisper==1.1.0
gTTS==2.5.4
pydub==0.25.1
pytest==8.3.4
httpx==0.28.1
